+++ src/Builder.ts
@@ -4,16 +4,17 @@
  */
 
 import isObject from './isObject';
-import { SupportedType, CheckerCallback, CustomCallback, OptimalOptions, Struct } from './types';
+import { SupportedType, CheckerCallback, CustomCallback, OptimalOptions } from './types';
 
 export interface Check {
   args: any[];
   callback: CheckerCallback;
 }
 
-export default class Builder<T> {
+export default class Builder<T, Struct extends object> {
   checks: Check[] = [];
 
+  // @ts-ignore Set before running checks
   currentStruct: Struct = {};
 
   defaultValue: T;
@@ -61,7 +62,7 @@ export default class Builder<T> {
   /**
    * Map a list of names that must be defined alongside this field.
    */
-  and(...keys: string[]): this {
+  and(...keys: (keyof Struct)[]): this {
     if (__DEV__) {
       this.invariant(keys.length > 0, 'AND requires a list of field names.');
     }
@@ -72,7 +73,7 @@ export default class Builder<T> {
   /**
    * Validate that all fields have been defined.
    */
-  checkAnd(path: string, value: any, otherKeys: string[]) {
+  checkAnd(path: string, value: T, otherKeys: (keyof Struct)[]) {
     if (__DEV__) {
       const keys = [this.key(path), ...otherKeys];
       const struct = this.currentStruct;
@@ -93,7 +94,7 @@ export default class Builder<T> {
   /**
    * Validate the type of value.
    */
-  checkType(path: string, value: any) {
+  checkType(path: string, value: T) {
     if (__DEV__) {
       switch (this.type) {
         case 'array':
@@ -122,7 +123,7 @@ export default class Builder<T> {
   /**
    * Set a callback to run custom logic.
    */
-  custom(callback: CustomCallback): this {
+  custom(callback: CustomCallback<Struct>): this {
     if (__DEV__) {
       this.invariant(
         typeof callback === 'function',
@@ -136,7 +137,7 @@ export default class Builder<T> {
   /**
    * Validate the value using a custom callback.
    */
-  checkCustom(path: string, value: any, callback: CustomCallback) {
+  checkCustom(path: string, value: T, callback: CustomCallback<Struct>) {
     if (__DEV__) {
       try {
         callback(value, this.currentStruct);
@@ -200,10 +201,10 @@ export default class Builder<T> {
   /**
    * Return the current key from a path.
    */
-  key(path: string): string {
+  key(path: string): keyof Struct {
     const index = path.lastIndexOf('.');
 
-    return index > 0 ? path.slice(index + 1) : path;
+    return (index > 0 ? path.slice(index + 1) : path) as keyof Struct;
   }
 
   /**
@@ -251,7 +252,7 @@ export default class Builder<T> {
   /**
    * Validate the value matches only the default value.
    */
-  checkOnly(path: string, value: any) {
+  checkOnly(path: string, value: T) {
     if (__DEV__) {
       this.invariant(
         value === this.defaultValue,
@@ -264,7 +265,7 @@ export default class Builder<T> {
   /**
    * Map a list of field names that must have at least 1 defined.
    */
-  or(...keys: string[]): this {
+  or(...keys: (keyof Struct)[]): this {
     if (__DEV__) {
       this.invariant(keys.length > 0, 'OR requires a list of field names.');
     }
@@ -275,7 +276,7 @@ export default class Builder<T> {
   /**
    * Validate that at least 1 field is defined.
    */
-  checkOr(path: string, value: any, otherKeys: string[]) {
+  checkOr(path: string, value: T, otherKeys: (keyof Struct)[]) {
     if (__DEV__) {
       const keys = [this.key(path), ...otherKeys];
       const struct = this.currentStruct;
@@ -302,7 +303,7 @@ export default class Builder<T> {
   /**
    * Run all validation checks that have been enqueued.
    */
-  runChecks(path: string, initialValue: any, struct: Struct, options: OptimalOptions = {}): any {
+  runChecks(path: string, initialValue: T, struct: Struct, options: OptimalOptions = {}): T | null {
     this.currentStruct = struct;
     this.options = options;
 
@@ -353,7 +354,7 @@ export default class Builder<T> {
   /**
    * Map a list of field names that must not be defined alongside this field.
    */
-  xor(...keys: string[]): this {
+  xor(...keys: (keyof Struct)[]): this {
     if (__DEV__) {
       this.invariant(keys.length > 0, 'XOR requires a list of field names.');
     }
@@ -364,7 +365,7 @@ export default class Builder<T> {
   /**
    * Validate that only 1 field is defined.
    */
-  checkXor(path: string, value: any, otherKeys: string[]) {
+  checkXor(path: string, value: T, otherKeys: (keyof Struct)[]) {
     if (__DEV__) {
       const keys = [this.key(path), ...otherKeys];
       const struct = this.currentStruct;
@@ -378,17 +379,17 @@ export default class Builder<T> {
   }
 }
 
-export function bool(defaultValue: boolean | null = false): Builder<boolean | null> {
-  return new Builder('boolean', defaultValue);
+export function bool<S extends object>(defaultValue: boolean | null = false) /* infer */ {
+  return new Builder<boolean | null, S>('boolean', defaultValue);
 }
 
-export function custom<T>(
-  callback: CustomCallback,
+export function custom<T, S extends object>(
+  callback: CustomCallback<S>,
   defaultValue: T | null = null,
-): Builder<T | null> {
-  return new Builder('custom', defaultValue).custom(callback);
+) /* infer */ {
+  return new Builder<T | null, S>('custom', defaultValue).custom(callback);
 }
 
-export function func(defaultValue: Function | null = null): Builder<Function | null> {
-  return new Builder('function', defaultValue).nullable();
+export function func<S extends object>(defaultValue: Function | null = null) /* infer */ {
+  return new Builder<Function | null, S>('function', defaultValue).nullable();
 }