+++ src/StringBuilder.ts
@@ -5,11 +5,11 @@
 
 import Builder from './Builder';
 
-function isString(value: any): boolean {
+function isString(value: any): value is string {
   return typeof value === 'string' && value !== '';
 }
 
-export default class StringBuilder extends Builder<string | null> {
+export default class StringBuilder<Struct extends object> extends Builder<string | null, Struct> {
   allowEmpty: boolean = false;
 
   constructor(defaultValue: string | null = '') {
@@ -29,7 +29,7 @@ export default class StringBuilder extends Builder<string | null> {
     return this.addCheck(this.checkContains, token, index);
   }
 
-  checkContains(path: string, value: any, token: string, index: number = 0) {
+  checkContains(path: string, value: string, token: string, index: number = 0) {
     if (__DEV__) {
       this.invariant(value.indexOf(token, index) >= 0, `String does not include "${token}".`, path);
     }
@@ -46,10 +46,10 @@ export default class StringBuilder extends Builder<string | null> {
     return this.addCheck(this.checkMatch, pattern);
   }
 
-  checkMatch(path: string, value: any, pattern: RegExp) {
+  checkMatch(path: string, value: string, pattern: RegExp) {
     if (__DEV__) {
       this.invariant(
-        value.match(pattern),
+        !!value.match(pattern),
         `String does not match pattern "${pattern.source}".`,
         path,
       );
@@ -64,7 +64,7 @@ export default class StringBuilder extends Builder<string | null> {
     return this;
   }
 
-  checkNotEmpty(path: string, value: any) {
+  checkNotEmpty(path: string, value: string) {
     if (__DEV__) {
       if (!this.allowEmpty) {
         this.invariant(isString(value), 'String cannot be empty.', path);
@@ -83,13 +83,13 @@ export default class StringBuilder extends Builder<string | null> {
     return this.addCheck(this.checkOneOf, list);
   }
 
-  checkOneOf(path: string, value: any, list: string[]) {
+  checkOneOf(path: string, value: string, list: string[]) {
     if (__DEV__) {
       this.invariant(list.indexOf(value) >= 0, `String must be one of: ${list.join(', ')}`, path);
     }
   }
 }
 
-export function string(defaultValue: string | null = ''): StringBuilder {
-  return new StringBuilder(defaultValue);
+export function string<S extends object>(defaultValue: string | null = '') /* infer */ {
+  return new StringBuilder<S>(defaultValue);
 }