+++ src/index.ts
@@ -5,9 +5,10 @@
 
 import optimal from './optimal';
 import Builder, { bool, custom, func } from './Builder';
-import CollectionBuilder, { array, object } from './CollectionBuilder';
+import ArrayBuilder, { array } from './ArrayBuilder';
 import InstanceBuilder, { instance, date, regex } from './InstanceBuilder';
 import NumberBuilder, { number } from './NumberBuilder';
+import ObjectBuilder, { object } from './ObjectBuilder';
 import ShapeBuilder, { shape } from './ShapeBuilder';
 import StringBuilder, { string } from './StringBuilder';
 import UnionBuilder, { union } from './UnionBuilder';
@@ -16,9 +17,10 @@ export { array, bool, custom, date, func, instance, number, object, regex, shape
 
 export {
   Builder,
-  CollectionBuilder,
+  ArrayBuilder,
   InstanceBuilder,
   NumberBuilder,
+  ObjectBuilder,
   ShapeBuilder,
   StringBuilder,
   UnionBuilder,