+++ tests/InstanceBuilder.test.ts
@@ -3,42 +3,48 @@ import InstanceBuilder, { instance, date, regex } from '../src/InstanceBuilder';
 describe('instance()', () => {
   class Foo {}
 
+  let builder: InstanceBuilder<Foo, { key: Foo | null }>;
+
+  beforeEach(() => {
+    builder = instance(Foo);
+  });
+
   describe('constructor()', () => {
     it('errors if a non-class is passed', () => {
       expect(() => {
-        // @ts-ignore
-        instance(123);
+        // @ts-ignore Test invalid type
+        builder = instance(123);
       }).toThrowErrorMatchingSnapshot();
     });
 
     it('errors if an object is passed', () => {
       expect(() => {
-        // @ts-ignore
-        instance({});
+        // @ts-ignore Test invalid type
+        builder = instance({});
       }).toThrowErrorMatchingSnapshot();
     });
 
     it('doesnt error if a class is passed', () => {
       expect(() => {
-        instance(Foo);
+        builder = instance(Foo);
       }).not.toThrowError('A class reference is required.');
     });
   });
 
   describe('runChecks()', () => {
     it('returns null for no data', () => {
-      expect(instance(Foo).runChecks('key', null, {})).toBeNull();
+      expect(builder.runChecks('key', null, { key: null })).toBeNull();
     });
 
     it('errors if a non-instance is passed', () => {
       expect(() => {
-        instance().runChecks('key', 'foo', {});
+        builder.runChecks('key', 'foo', { key: null });
       }).toThrowErrorMatchingSnapshot();
     });
 
     it('errors if an object is passed when a class instance is required', () => {
       expect(() => {
-        instance().runChecks('key', {}, {});
+        builder.runChecks('key', {}, { key: null });
       }).toThrowErrorMatchingSnapshot();
     });