+++ tests/StringBuilder.test.ts
@@ -1,7 +1,7 @@
 import StringBuilder, { string } from '../src/StringBuilder';
 
 describe('StringBuilder', () => {
-  let builder: StringBuilder;
+  let builder: StringBuilder<{ key: string }>;
 
   beforeEach(() => {
     builder = string('foo');
@@ -16,7 +16,8 @@ describe('StringBuilder', () => {
   describe('runChecks()', () => {
     it('errors if a non-string value is used', () => {
       expect(() => {
-        builder.runChecks('key', 123, {});
+        // @ts-ignore Testing wrong type
+        builder.runChecks('key', 123, { key: 123 });
       }).toThrowErrorMatchingSnapshot();
     });
   });
@@ -24,7 +25,7 @@ describe('StringBuilder', () => {
   describe('contains()', () => {
     it('errors if token is not string', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.contains(123);
       }).toThrowErrorMatchingSnapshot();
     });
@@ -62,7 +63,7 @@ describe('StringBuilder', () => {
   describe('match()', () => {
     it('errors if pattern is not a regex', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.match(123);
       }).toThrowErrorMatchingSnapshot();
     });
@@ -116,7 +117,7 @@ describe('StringBuilder', () => {
   describe('oneOf()', () => {
     it('errors if not an array', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.oneOf(123);
       }).toThrowErrorMatchingSnapshot();
     });
@@ -129,7 +130,7 @@ describe('StringBuilder', () => {
 
     it('errors if array contains a non-string', () => {
       expect(() => {
-        // @ts-ignore
+        // @ts-ignore Testing wrong type
         builder.oneOf(['foo', 123]);
       }).toThrowErrorMatchingSnapshot();
     });